// frontend/client/models/test-runs-trend.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Initry backend
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface TestRunsTrend
 */
export interface TestRunsTrend {
    /**
     * 
     * @type {Array<number>}
     * @memberof TestRunsTrend
     */
    'passed': Array<number>;
    /**
     * 
     * @type {Array<number>}
     * @memberof TestRunsTrend
     */
    'failed': Array<number>;
    /**
     * 
     * @type {Array<number>}
     * @memberof TestRunsTrend
     */
    'skipped': Array<number>;
}